Projet_AIF/requirements.txt
torch==2.0.1
torchvision==0.15.2
tensorboard==2.14.0
Flask==2.3.2
Pillow==10.0.0
pandas==2.0.3
numpy==1.24.4
matplotlib==3.7.2
gradio==3.39.0
